Adding classifyer but not for this touchsim data set
This script requires adjustments to work with the touchsim outputs. But in principle shows how to classify with the function. 
For experiment classy with young and compare to bahvioural data or for each spific afferent to check which afferent is more improtant for this task or to see if any of the manipulations has any effects e.g. youngs modulus and afferent density etc.